# Conflicts:
#	src/main/java/buisnessLogicLayer/ResponseHandler.java
#	src/main/java/dataAccessLayer/repositories/OrderRepository.java
#	src/main/java/shared/RequestMessage.java
#	src/main/java/shared/TypeOfMessage.java
#	test.iml

+++ src/main/java/client/view/UserInterface.java
@@ -117,8 +117,8 @@ public class UserInterface {
         return EnumHandler.getType(scanner.nextInt());
     }
 
-    public int[] getPriceRange() {
-        int[] range = new int[2];
+    public double[] getPriceRange() {
+        double[] range = new double[2];
         System.out.println("Type min prize");
         range[0] = scanner.nextInt();
         System.out.println("Type max prize");
@@ -164,7 +164,6 @@ public class UserInterface {
             if (selectedId == itemId) {
                 return object;
             }
-
         }
         System.out.println("Invalid product id");
         return null;
